+++ drools-core/src/main/java/org/drools/core/time/TimerService.java
@@ -50,6 +50,6 @@ public interface TimerService extends SchedulerService {
      * This method may return null for some TimerService implementations that do not want the overhead of maintain this.
      * @return
      */
-    public Collection<TimerJobInstance> getTimerJobInstances(int id);
+    public Collection<TimerJobInstance> getTimerJobInstances(long id);
     
 }
